Move Importmap to its own module and refactor
Extract Importmap implementation into new imap.js, simplify importmap.mjs to re-export JSON, and update imports across the codebase (index.mjs, html.mjs, importmap.test.js). Add Importmap.importFromScript() (browser-only), change constructor to avoid defaulting imports/scopes, and update getScript/getIntegrity to support named scripts. Update consts.js regex to capture script name and nonce, add tests for HTML parsing and relative resolution, and use dynamic imports for node:fs/path where needed to improve browser compatibility. Also bump package version to 1.9.0 and add a changelog entry.
