+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@shgysk8zer0/importmap",
-  "version": "1.8.20",
+  "version": "1.9.0",
   "engines": {
     "node": ">=20.10.0"
   },
